Ensure timers don't overlap and are cleaned up
Fixes a bug where if a sync was begun, ended, and a new sync begun all
inside five minutes, the original timer from the first sync would
trigger the new sync to end early.

The fix is to use unique timers for all situations, track them, and stop
them manually when no longer in use. This should also be somewhat more
memory-efficient, as it allows the threads to stop and get cleaned up.